+++ NEWS.md
@@ -13,6 +13,7 @@ version 0.2.3.9000
 * User can now specify `modulePath` as a character vector, e.g., `simInit(..., paths = list(modulePath = c(".", "test")))`. This means that a user can organize the modules in different locations.
 * `modulePath` now has a new argument, `module`, where user can specify (a) specific module(s)'s path. Modifications were implemented to `dataPath` to utilize this new feature
 * `simInit` and `spades` now call `setPaths(paths)` or `setPaths(sim$paths)`, unsetting them `on.exit` internally to make the paths used for functions e.g., `reproducible::Cache` to use the correct path
+* under-the-hood speed improvements for the DES (about 20% faster) -- 38 microseconds per event under ideal conditions
 
 ## bug fixes